Moving CuTensorNetHandle to the general.py at the root of the module
