+++ pytket/extensions/cutensornet/general_state/tensor_network_state.py
@@ -11,8 +11,7 @@ import numpy as np
 from sympy import Expr  # type: ignore
 from numpy.typing import NDArray
 from pytket.circuit import Circuit
-from pytket.extensions.cutensornet.general import set_logger
-from pytket.extensions.cutensornet.structured_state import CuTensorNetHandle
+from pytket.extensions.cutensornet.general import CuTensorNetHandle, set_logger
 from pytket.utils.operators import QubitPauliOperator
 
 try:
