+++ pytket/extensions/cutensornet/structured_state/mps.py
@@ -32,9 +32,9 @@ except ImportError:
 from pytket.circuit import Command, Op, OpType, Qubit
 from pytket.pauli import Pauli, QubitPauliString
 
-from pytket.extensions.cutensornet.general import set_logger
+from pytket.extensions.cutensornet.general import CuTensorNetHandle, set_logger
 
-from .general import CuTensorNetHandle, Config, StructuredState, Tensor
+from .general import Config, StructuredState, Tensor
 
 
 class DirMPS(Enum):
